requirements.txt
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
anyio==3.6.1
attrs==21.4.0
bcrypt==3.2.2
bidict==0.23.1
billiard==4.2.0
cachelib==0.13.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.14
dnspython==2.6.1
email-validator==1.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-RESTful==0.3.9
flask-restx==0.5.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.6.0
kombu==5.3.7
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.21.3
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.28.2
mdurl==0.1.2
mongoengine==0.24.1
mongomock==4.1.2
oauthlib==2.1.0
ordered-set==4.1.0
packaging==24.1
passlib==1.7.4
phonenumbers==8.13.45
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.9.1
Pygments==2.18.0
PyJWT==2.4.0
pymongo==4.7.3
PyMySQL==1.0.2
pymysql-pool==0.3.7
pyrsistent==0.18.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-decouple==3.6
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2022.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==1.1.0
rich==13.8.1
sentinels==1.0.0
setuptools==74.1.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.20.3
typing_extensions==4.2.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.1.2
wrapt==1.16.0
wsproto==1.2.0
redis
